move useEffect to store access token to Heat.tsx
